refactor(blorktools): custom animation package
