+++ packages/blorktools/src/asset_debugger/util/preview/threejs-preview-util.js
@@ -1,46 +1,24 @@
-import * as THREE from 'three';
-import { animatePreview, previewAnimationId, resetLastAnimationFrameTime, resetPreviewAnimationId } from '../animation/playback/animation-preview-util';
+import { infoPanel, resetInfoPanel } from "../../modals/html-editor-modal/mesh-info-panel-util";
+import { logPreviewError } from "../log-util";
+import { cleanupCSS3D } from "../scene/css3d-scene-util";
+import { cleanupThreeJsScene, setupThreeJsScene } from "../scene/threejs-scene-util";
+import { setIsPreviewActive, setIsPreviewAnimationPaused, setLastTextureUpdateTime } from "../state/animation-state";
+import { reverseAnimationFrameId } from "../state/css3d-state";
 import { 
-    isPreviewActive,
-    setIsPreviewActive, 
-    setIsPreviewAnimationPaused, 
-    setLastTextureUpdateTime,
-} from '../state/animation-state';
-import { getState } from '../../scene/state';
-import { showStatus } from '../../modals/html-editor-modal/html-editor-modal';
-import { createMeshInfoPanel, infoPanel, resetInfoPanel } from '../../modals/html-editor-modal/mesh-info-panel-util';
-import { resetAnimationState, reverseAnimationFrameId } from '../state/css3d-state';
-import { createTextureFromIframe } from '../animation/render/iframe2texture-render';
-import { cleanupThreeJsScene, setupThreeJsScene } from '../scene/threejs-scene-util';
-
-let pendingTextureUpdate = false;
-export let previewPlane;
-export let animationPreviewScene, animationPreviewCamera, animationPreviewRenderer;
-export let animationCss3dScene, animationCss3dRenderer, animationCss3dObject;
-export let frameBuffer = [];
-export let previewRenderTarget = null;
-
-/**
- * Initialize Three.js for HTML preview
- * @param {HTMLElement} container - The container element for the Three.js canvas
- * @param {HTMLIFrameElement} iframe - The iframe containing the HTML to render as texture
- * @param {number} currentMeshId - The ID of the current mesh
- * @param {boolean} createInfoPanel - Whether to create the info panel
- */
-export function initThreeJsPreview(container, iframe, currentMeshId, createInfoPanel = true) {
-    try {
-        // We already have THREE imported at the top of the file
-        console.log('Using imported Three.js module');
-
-        // Only need to load html2canvas
-        loadHtml2Canvas(() => {
-            setupThreeJsScene(container, iframe, currentMeshId, createInfoPanel);
-        });
-    } catch (error) {
-        console.error('Error initializing Three.js preview:', error);
-        logPreviewError(`Three.js initialization error: ${error.message}`);
-    }
-}
+    animationCss3dObject,
+    animationCss3dRenderer,
+    animationCss3dScene,
+    animationPreviewCamera,
+    animationPreviewRenderer,
+    animationPreviewScene,
+    frameBuffer,
+    pendingTextureUpdate, 
+    previewPlane, 
+    previewRenderTarget, 
+    resetThreeJsState, 
+    setPendingTextureUpdate 
+} from "../state/threejs-state";
+import { previewAnimationId, resetLastAnimationFrameTime, resetPreviewAnimationId } from "./animation-preview-util";
 
 /**
  * Clean up Three.js preview resources using the generic cleanup utility
@@ -54,7 +32,7 @@ export function cleanupThreeJsPreview() {
     
     // Set pending operations to false
     if (pendingTextureUpdate) {
-        pendingTextureUpdate = false;
+        setPendingTextureUpdate(false);
     }
 
     // Collect DOM elements to clean up
@@ -113,15 +91,7 @@ export function cleanupThreeJsPreview() {
     });
 
     // Reset global variables
-    previewPlane = null;
-    animationPreviewScene = null;
-    animationPreviewCamera = null;
-    animationPreviewRenderer = null;
-    animationCss3dScene = null;
-    animationCss3dRenderer = null;
-    animationCss3dObject = null;
-    previewRenderTarget = null;
-    frameBuffer = [];
+    resetThreeJsState();
 
     // Reset debug flag and global functions
     window._css3dDebugLogged = false;
@@ -132,13 +102,15 @@ export function cleanupThreeJsPreview() {
     // Reset animation state
     setIsPreviewAnimationPaused(false);
     setLastTextureUpdateTime(0);
-    pendingTextureUpdate = false;
+    // TODO Move this variable to animation state object
+    setPendingTextureUpdate(false);
     resetLastAnimationFrameTime();
     resetPreviewAnimationId();
 
     console.log('Three.js and CSS3D resources cleaned up');
 }
 
+// TODO Rename to seomthing better and make it not coupled to specific panel
 /**
  * Clean up info panel resources
  * Removes the info panel from the DOM and resets the infoPanel reference
@@ -160,7 +132,7 @@ function cleanupInfoPanel() {
 /**
  * Handle window resize for the Three.js preview
  */
-export function onPreviewResize() {
+function onPreviewResize() {
     const container = animationPreviewRenderer.domElement.parentElement;
     if (!container) return;
     
@@ -178,19 +150,27 @@ export function onPreviewResize() {
 }
 
 /**
- * Set preview render target
- * @param {THREE.WebGLRenderTarget} incomingValue - The new value to set
+ * Initialize Three.js for HTML preview
+ * @param {HTMLElement} container - The container element for the Three.js canvas
+ * @param {HTMLIFrameElement} iframe - The iframe containing the HTML to render as texture
+ * @param {number} currentMeshId - The ID of the current mesh
+ * @param {boolean} createInfoPanel - Whether to create the info panel
  */
-export function setPreviewRenderTarget(incomingValue) {
-    previewRenderTarget = incomingValue;
-}
+export function initThreeJsPreview(container, iframe, currentMeshId, createInfoPanel = true) {
+    // Handle window resize
+    window.addEventListener('resize', onPreviewResize);
+    try {
+        // We already have THREE imported at the top of the file
+        console.log('Using imported Three.js module');
 
-/**
- * Set aniamtion preview camera
- * @param {THREE.Camera} incomingValue - The new value to set
- */
-export function setAnimationPreviewCamera(incomingValue) {
-    animationPreviewCamera = incomingValue;
+        // Only need to load html2canvas
+        loadHtml2Canvas(() => {
+            setupThreeJsScene(container, iframe, currentMeshId, createInfoPanel);
+        });
+    } catch (error) {
+        console.error('Error initializing Three.js preview:', error);
+        logPreviewError(`Three.js initialization error: ${error.message}`);
+    }
 }
 
 /**
@@ -224,72 +204,4 @@ function loadHtml2Canvas(callback) {
         console.error('Failed to load html2canvas:', error);
     };
     document.head.appendChild(script);
-}
-
-function cleanupCSS3D(targetElement = null) {
-   resetAnimationState();
-   
-   if (reverseAnimationFrameId) {
-       cancelAnimationFrame(reverseAnimationFrameId);
-       resetReverseAnimationFrameId();
-   }
-   
-   const iframe = targetElement || document.getElementById('css3d-panel-iframe');
-   if (iframe) {
-       if (iframe.restartTimer) {
-           clearInterval(iframe.restartTimer);
-           iframe.restartTimer = null;
-       }
-       
-       if (iframe._animationStartHandler && iframe.contentDocument) {
-           iframe.contentDocument.removeEventListener('animationstart', iframe._animationStartHandler);
-           iframe._animationStartHandler = null;
-       }
-       
-       if (iframe._transitionStartHandler && iframe.contentDocument) {
-           iframe.contentDocument.removeEventListener('transitionstart', iframe._transitionStartHandler);
-           iframe._transitionStartHandler = null;
-       }
-       
-       if (iframe.mutationObserver) {
-           iframe.mutationObserver.disconnect();
-           iframe.mutationObserver = null;
-       }
-   }
-}
-
-/**
- * Set aniamtion css 3d scene
- * @param {THREE.Scene} incomingValue - The new value to set
- */
-export function setAnimationCss3dScene(incomingValue) {
-    animationCss3dScene = incomingValue;
-}
-
-/**
- * Set animation css 3d renderer
- * @param {THREE.WebGLRenderer} incomingValue - The new value to set
- */
-export function setAnimationCss3dRenderer(incomingValue) {
-    animationCss3dRenderer = incomingValue;
-}
-
-/**
- * Set animation css 3d object
- * @param {THREE.Object3D} incomingValue - The new value to set
- */
-export function setAnimationCss3dObject(incomingValue) {
-    animationCss3dObject = incomingValue;
-}
-
-export function setAnimationPreviewScene(incomingValue) {
-    animationPreviewScene = incomingValue;
-}
-
-export function setAnimationPreviewRenderer(incomingValue) {
-    animationPreviewRenderer = incomingValue;
-}
-
-export function setPreviewPlane(incomingValue) {
-    previewPlane = incomingValue;
 }
